+++ gDRcore/R/data_type.R
@@ -1,15 +1,12 @@
 #' Identify type of data
 #'
 #' @param df data.frame of raw drug response data containing both treated and untreated values
-#' @param cotreatment_conc integer of maximum number of concentration of co-treatment
-#' to classify as cotreatment data type.
-#' Defaults to \code{4}.
 #' @param codilution_conc integer of maximum number of concentration ratio of co-treatment
 #' to classify as codilution data type.
 #' Defaults to \code{2}.
 #' @param matrix_conc integer of minimum number of concentration pairs of co-treatment
-#' to classify as matrix data type.
-#' Defaults to \code{4}.
+#' to classify as co-treatment or matrix data type.
+#' Defaults to \code{1}.
 #'
 #' @return data.frame of raw drug response data with additional column `type` with the info of
 #' data type for a given row of data.frame
@@ -17,9 +14,8 @@
 #'
 #' @author Bartosz Czech <bartosz.czech@@contractors.roche.com>
 identify_data_type <- function(df,
-                               cotreatment_conc = 4,
                                codilution_conc = 2,
-                               matrix_conc = 4
+                               matrix_conc = 1      # forces any co-treatment as a matrix data model
                                ) {
   
   # find the pairs of drugs with relevant metadata
@@ -81,12 +77,8 @@ identify_data_type <- function(df,
     type <- 
       if (length(conc_ratio) <= codilution_conc) {
         "co-dilution"
-      } else if (n_conc_pairs == length(conc_1) * length(conc_2) & length(conc_2) >= matrix_conc) {
-        "matrix"
-      } else if (length(conc_2) < cotreatment_conc) {
-        "cotreatment"
       } else {
-        "other"
+        "matrix"
       }
     df[matching_idx, "type"]  <- ifelse(is.na(df[matching_idx, "type"]), type, df[matching_idx, "type"])
   }
@@ -113,6 +105,7 @@ split_raw_data <- function(df,
   drug_ids <- drug_ids[which(drug_ids %in% names(df))]
   codrug_ids <- drug_ids[grep("[0-9]", names(drug_ids))]
   conc_idx <- drug_ids[grep("concentration", names(drug_ids))]
+  codrug_drug_id <- setdiff(codrug_ids, conc_idx)
   
   cl <- gDRutils::get_env_identifiers("cellline")
   df_list <- split(df, df[[type_col]])
@@ -133,7 +126,10 @@ split_raw_data <- function(df,
       cotrt_matching <- rbind(unique_cotrt, unique_cotrt_ctrl)
       df_merged <- rbind(df_list[[x]], merge(cotrt_matching, control))
       if (x == "matrix") {
-        rbind(df_merged, df_list[["single-agent"]])
+        matrix_data <- rbind(df_merged, df_list[["single-agent"]])
+        matrix_data[, conc_idx][is.na(matrix_data[, conc_idx])] <- 0
+        matrix_data[, codrug_drug_id][is.na(matrix_data[, codrug_drug_id])] <- untreated_tag[1]
+        matrix_data
       } else {
         df_merged
       }