+++ gDRcore/R/fit_SE.combinations.R
@@ -65,6 +65,7 @@ fit_SE.combinations <- function(se,
       smooth_mx <- hsa_excess <- bliss_excess <- isobolograms <- metrics <- 
         bliss_score[, c("row_id", "col_id")] <- hsa_score[, c("row_id", "col_id")] <-
         CIScore_50[, c("row_id", "col_id")] <- CIScore_80[, c("row_id", "col_id")] <-
+        all_iso_points <-
         data.frame(row_id = i, col_id = j)
       return(list(bliss_excess = bliss_excess,
            hsa_excess = hsa_excess,
@@ -106,8 +107,14 @@ fit_SE.combinations <- function(se,
       # fit by column: the series in the primary identifier, the cotrt is the secondary one
       col_fittings <- fit_combo_cotreatments(avg_combo,
                                              series_id = id, cotrt_id = id2, metric)
-      col_fittings <- col_fittings[!is.na(col_fittings$fit_type), ]
-
+      col_fittings <- if (all(is.na(col_fittings$fit_type))) {
+        x <- col_fittings[1, ]
+        x[, "cotrt_value"] <- NA
+        x
+      } else {
+        col_fittings[!is.na(col_fittings$fit_type), ]
+      }
+    
       # fit by row (flipped): the series in the secondary identifier, the cotrt is the primary one
       row_fittings <- fit_combo_cotreatments(avg_combo,
                                              series_id = id2, cotrt_id = id, metric)
@@ -138,8 +145,11 @@ fit_SE.combinations <- function(se,
       # we need it to distinguish which rows are col_fittings/row_fittings and codilution_fitting
       metrics_merged$source <- rep(metrics_names, vapply(mget(metrics_names), nrow, numeric(1)))
       # remove degenerated fits
-      metrics_merged <- metrics_merged[metrics_merged$N_conc > 1, ]
-
+      metrics_merged <-
+        metrics_merged[!(metrics_merged$fit_type %in% c("DRCInvalidFitResult", "DRCTooFewPointsToFit")), ]
+      if (nrow(metrics_merged) == 0) {
+        metrics_merged[1, ] <- NA
+      }
       keep <- intersect(colnames(complete), c(metric, "row_values", "col_values", "codil_values"))
       mat <- as.matrix(complete[, keep])
       complete$average <- rowMeans(mat, na.rm = TRUE)
@@ -174,7 +184,8 @@ fit_SE.combinations <- function(se,
                         series_identifiers = c(id, id2), normalization_type = metric
                         )
       } else {
-        NULL
+        list(df_all_iso_points = data.frame(row_id = NA, col_id = NA),
+             df_all_iso_curves = data.frame(row_id = NA, col_id = NA))
       }
 
       # average the top 10-percentile excess to get a single value for the excess