+++ gDRcore/tests/testthat/test-finalMAE_combo_matrix.R
@@ -2,6 +2,6 @@ test_that("combo_matrix: test_synthetic_data", {
   data <- "finalMAE_combo_matrix.RDS"
   original <- get_synthetic_data(data)
   mae <- purrr::quietly(gDRtestData::generateComboMatrix)(cell_lines, drugs, e_inf, ec50, hill_coef, FALSE)
-  expect_length(mae$warnings, 6)
+  expect_length(mae$warnings, 4)
   test_synthetic_data(original, mae$result, data)
 })